@@ -6,23 +6,24 @@
   "author": "mizdra <pp.mizdra@gmail.com>",
   "license": "MIT",
   "private": false,
+  "type": "module",
   "scripts": {
     "build": "tsc -p tsconfig.src.json",
-    "dev": "tsc-watch -p tsconfig.src.json --onSuccess 'bin/eslint-interactive fixtures --rulesdir fixtures/rules --ext .js,.jsx,.mjs'",
+    "dev": "tsc-watch -p tsconfig.src.json --onSuccess 'bin/eslint-interactive.js fixtures --rulesdir fixtures/rules --ext .js,.jsx,.mjs'",
     "lint": "run-s -c lint:*",
     "lint:tsc": "run-s -c lint:tsc:*",
     "lint:tsc:src": "tsc -p tsconfig.src.json --noEmit",
     "lint:tsc:test": "tsc -p tsconfig.test.json --noEmit",
     "lint:eslint": "eslint --ignore-pattern '/fixtures' .",
     "lint:prettier": "prettier --check .",
-    "test": "jest --colors"
+    "test": "NODE_OPTIONS=--experimental-vm-modules jest --colors"
   },
   "prettier": "@mizdra/prettier-config-mizdra",
   "renovate": "github>mizdra/renovate-config-mizdra",
   "packageManager": "yarn@1.22.15",
   "devDependencies": {
-    "@jest/types": "^27.1.0",
-    "@mizdra/eslint-config-mizdra": "^1.0.0",
+    "@jest/types": "^27.4.2",
+    "@mizdra/eslint-config-mizdra": "^1.1.0",
     "@mizdra/prettier-config-mizdra": "^0.4.0",
     "@types/cli-table": "^0.3.0",
     "@types/eslint": "^8.2.0",
@@ -34,21 +35,22 @@
     "@types/yargs": "^16.0.4",
     "@typescript-eslint/eslint-plugin": "^5.2.0",
     "@typescript-eslint/parser": "^5.2.0",
+    "enhanced-resolve": "^5.8.3",
     "eslint": "^8.1.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.25.2",
-    "jest": "^27.1.0",
+    "jest": "^27.4.5",
     "jest-mock-process": "^1.4.1",
     "npm-run-all": "^4.1.5",
     "prettier": "2.4.1",
-    "strip-ansi": "^6.0.0",
-    "ts-jest": "^27.0.5",
+    "strip-ansi": "^7.0.1",
+    "ts-jest": "^27.1.2",
     "tsc-watch": "^4.2.9",
-    "typescript": "^4.1.2"
+    "typescript": "^4.6.0-dev.20211228"
   },
   "dependencies": {
-    "boxen": "^5.1.2",
-    "chalk": "^4.1.2",
+    "boxen": "^6.2.1",
+    "chalk": "^5.0.0",
     "cli-table": "^0.3.11",
     "comlink": "^4.3.1",
     "enquirer": "^2.3.6",
@@ -56,23 +58,23 @@
     "estraverse": "^5.3.0",
     "is-installed-globally": "^0.4.0",
     "node-pager": "^0.3.6",
-    "ora": "^5.1.0",
+    "ora": "^6.0.1",
     "source-map-support": "^0.5.20",
-    "terminal-link": "^2.1.1",
+    "terminal-link": "^3.0.0",
     "yargs": "^16.2.0"
   },
   "peerDependencies": {
     "eslint": "^7.0.0 || ^8.0.0"
   },
   "engines": {
-    "node": ">=14.0"
+    "node": "^14.13.1 || >=16.0.0"
   },
   "publishConfig": {
     "access": "public",
     "registry": "https://registry.npmjs.org/"
   },
   "bin": {
-    "eslint-interactive": "bin/eslint-interactive"
+    "eslint-interactive": "bin/eslint-interactive.js"
   },
   "keywords": [
     "eslint"
