@@ -21,7 +21,7 @@ module.exports = {
   overrides: [
     // for typescript
     {
-      files: ['*.ts', '*.tsx'],
+      files: ['*.ts', '*.tsx', '*.cts', '*.mts'],
       parserOptions: {
         project: ['./tsconfig.src.json', './tsconfig.test.json'],
       },
@@ -32,11 +32,14 @@ module.exports = {
         '@typescript-eslint/no-unsafe-return': 'off',
         // 煩すぎるので off
         '@typescript-eslint/no-unsafe-argument': 'off',
+        // disable because this rule do not support ESM in TypeScript.
+        // ref: https://github.com/import-js/eslint-plugin-import/issues/2170
+        'import/no-unresolved': 'off',
       },
     },
     // for test
     {
-      files: ['test/**/*.{ts,tsx}'],
+      files: ['test/**/*.{ts,tsx,cts,mts}'],
       env: {
         jest: true,
       },
