@@ -2,8 +2,8 @@ import chalk from 'chalk';
 import { Remote } from 'comlink';
 import { ESLint } from 'eslint';
 import ora from 'ora';
-import { promptToInputDescription } from '../cli/prompt';
-import { SerializableCore } from '../core-worker';
+import { promptToInputDescription } from '../cli/prompt.js';
+import { SerializableCore } from '../core-worker.js';
 
 export async function doDisablePerLineAction(
   core: Remote<SerializableCore>,
