@@ -1,7 +1,7 @@
 import chalk from 'chalk';
 import { Remote } from 'comlink';
 import ora from 'ora';
-import { SerializableCore } from '../core-worker';
+import { SerializableCore } from '../core-worker.js';
 
 export async function doFixAction(core: Remote<SerializableCore>, selectedRuleIds: string[]) {
   const fixingSpinner = ora('Fixing...').start();
