@@ -1,7 +1,7 @@
 import { Remote } from 'comlink';
 import { ESLint } from 'eslint';
-import { promptToInputDisplayMode } from '../cli/prompt';
-import { SerializableCore } from '../core-worker';
+import { promptToInputDisplayMode } from '../cli/prompt.js';
+import { SerializableCore } from '../core-worker.js';
 
 export async function doPrintDetailsOfResultsAction(
   core: Remote<SerializableCore>,
