@@ -2,12 +2,12 @@ import { tmpdir } from 'os';
 import { join } from 'path';
 import { ESLint } from 'eslint';
 import pager from 'node-pager';
-import { format } from './formatter';
-import transformRule, { TransformRuleOption } from './rules/transform';
-import { SuggestionFilter } from './transforms/apply-suggestions';
-import { FixableMaker } from './transforms/make-fixable-and-fix';
-import { Config, DisplayMode, Transform } from './types';
-import { filterResultsByRuleId, scanUsedPluginsFromResults } from './util/eslint';
+import { format } from './formatter/index.js';
+import transformRule, { TransformRuleOption } from './rules/transform.js';
+import { SuggestionFilter } from './transforms/apply-suggestions.js';
+import { FixableMaker } from './transforms/make-fixable-and-fix.js';
+import { Config, DisplayMode, Transform } from './types.js';
+import { filterResultsByRuleId, scanUsedPluginsFromResults } from './util/eslint.js';
 
 /**
  * The core of eslint-interactive.
