@@ -1,6 +1,6 @@
 import chalk from 'chalk';
 import { ESLint } from 'eslint';
-import { ERROR_COLOR, FAILED_COLOR, WARNING_COLOR } from './colors';
+import { ERROR_COLOR, FAILED_COLOR, WARNING_COLOR } from './colors.js';
 
 function pluralize(word: string, count: number): string {
   return count > 1 ? `${word}s` : word;
