@@ -1,6 +1,6 @@
 import { ESLint } from 'eslint';
-import { formatByFiles } from './format-by-files';
-import { formatByRules } from './format-by-rules';
+import { formatByFiles } from './format-by-files.js';
+import { formatByRules } from './format-by-rules.js';
 
 export const format: ESLint.Formatter['format'] = (results, data) => {
   return formatByFiles(results) + '\n' + formatByRules(results, data);
