@@ -1,10 +1,10 @@
 import { Linter, Rule, SourceCode } from 'eslint';
-import { SelectActionArgs } from './scenes/select-action';
-import { SelectRuleIdsArgs } from './scenes/select-rule-ids';
-import { TransformToApplySuggestionsArgs } from './transforms/apply-suggestions';
-import { TransformToDisablePerFileArgs } from './transforms/disable-per-file';
-import { TransformToDisablePerLineArgs } from './transforms/disable-per-line';
-import { TransformToMakeFixableAndFixArgs } from './transforms/make-fixable-and-fix';
+import { SelectActionArgs } from './scenes/select-action.js';
+import { SelectRuleIdsArgs } from './scenes/select-rule-ids.js';
+import { TransformToApplySuggestionsArgs } from './transforms/apply-suggestions.js';
+import { TransformToDisablePerFileArgs } from './transforms/disable-per-file.js';
+import { TransformToDisablePerLineArgs } from './transforms/disable-per-line.js';
+import { TransformToMakeFixableAndFixArgs } from './transforms/make-fixable-and-fix.js';
 
 /**
  * The type that indicates what to do with the problems of selected rules.
