@@ -1,11 +1,13 @@
 import { exec } from 'child_process';
-import { join } from 'path';
+import { dirname, join } from 'path';
+import { fileURLToPath } from 'url';
 import { promisify } from 'util';
 import { ESLint } from 'eslint';
 import { mockConsoleLog } from 'jest-mock-process';
-import { Core } from '../src/core';
+import { Core } from '../src/core.js';
 
 const execPromise = promisify(exec);
+const cwd = join(dirname(fileURLToPath(import.meta.url)), '..');
 
 /**
  * Returns a string containing the stitched together contents of the file modified by transform.
@@ -13,7 +15,7 @@ const execPromise = promisify(exec);
  */
 async function getSnapshotOfChangedFiles(): Promise<string> {
   const { stdout } = await execPromise(`diff -qr fixtures fixtures-tmp | cut -d " " -f 4 | xargs tail -n +1`, {
-    cwd: join(__dirname, '..'),
+    cwd,
   });
   return stdout.toString();
 }
@@ -32,11 +34,11 @@ function normalize(results: ESLint.LintResult[]): ESLint.LintResult[] {
 }
 
 beforeEach(async () => {
-  await execPromise(`rm -rf fixtures-tmp && cp -r fixtures fixtures-tmp`, { cwd: join(__dirname, '..') });
+  await execPromise(`rm -rf fixtures-tmp && cp -r fixtures fixtures-tmp`, { cwd });
 });
 
 afterEach(async () => {
-  await execPromise(`rm -rf fixtures-tmp`, { cwd: join(__dirname, '..') });
+  await execPromise(`rm -rf fixtures-tmp`, { cwd });
 });
 
 describe('Core', () => {
@@ -45,7 +47,7 @@ describe('Core', () => {
     rulePaths: ['fixtures-tmp/rules'],
     extensions: ['.js', '.jsx', '.mjs'],
     formatterName: 'codeframe',
-    cwd: join(__dirname, '..'),
+    cwd,
   });
   test('lint', async () => {
     const results = await core.lint();
