@@ -56,6 +56,5 @@ const rule: Rule.RuleModule = {
   },
 };
 
-module.exports = rule; // for ESLint's Node.js API
 // eslint-disable-next-line import/no-default-export
-export default rule; // for test
+export default rule;
