@@ -1,5 +1,5 @@
-import { createTransformToDisablePerFile } from '../../src/transforms/disable-per-file';
-import { TransformTester } from '../test-util/transform-tester';
+import { createTransformToDisablePerFile } from '../../src/transforms/disable-per-file.js';
+import { TransformTester } from '../test-util/transform-tester.js';
 
 const tester = new TransformTester(
   createTransformToDisablePerFile,
